Refactor manifest creation scripts to set script directory and update execution path
